refactor(ui): colocate caching, projects, and guardrails-monitor into _components/
Continues the colocation follow-up (after api-reference #32690 and the
memory/budgets/access-groups batch #32694), applying the same convention to
three more self-contained segments: rename each `<segment>/components/` to
`_components/` and repoint the one `page.tsx` import.

Each folder is imported only by its own page (verified zero external importers
across src, tests, and e2e_tests, and no absolute self-imports inside the
folder), so this is a pure relocation with no logic change. The `@/components`
imports in these segments are all genuinely shared (networking, shared/*,
common_components, llm_calls, molecules) and stay put; nothing is drained.

Grandfathered lint suppressions under the three folders (14 entries total:
caching 4, projects 4, guardrails-monitor 6) are re-keyed to the new paths with
counts unchanged.
