+++ ui/litellm-dashboard/eslint-suppressions.json
@@ -29,7 +29,7 @@
       "count": 1
     }
   },
-  "src/app/(dashboard)/caching/components/cache_dashboard.tsx": {
+  "src/app/(dashboard)/caching/_components/cache_dashboard.tsx": {
     "no-restricted-imports": {
       "count": 1
     },
@@ -40,17 +40,17 @@
       "count": 2
     }
   },
-  "src/app/(dashboard)/caching/components/cache_health.tsx": {
+  "src/app/(dashboard)/caching/_components/cache_health.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/caching/components/cache_settings/RedisTypeSelector.tsx": {
+  "src/app/(dashboard)/caching/_components/cache_settings/RedisTypeSelector.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/caching/components/cache_settings/index.tsx": {
+  "src/app/(dashboard)/caching/_components/cache_settings/index.tsx": {
     "no-restricted-imports": {
       "count": 1
     },
@@ -136,32 +136,32 @@
       "count": 2
     }
   },
-  "src/app/(dashboard)/guardrails-monitor/components/EvaluationSettingsModal.tsx": {
+  "src/app/(dashboard)/guardrails-monitor/_components/EvaluationSettingsModal.tsx": {
     "react-hooks/set-state-in-effect": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/guardrails-monitor/components/GuardrailDetail.tsx": {
+  "src/app/(dashboard)/guardrails-monitor/_components/GuardrailDetail.tsx": {
     "no-nested-ternary": {
       "count": 3
     }
   },
-  "src/app/(dashboard)/guardrails-monitor/components/GuardrailsMonitorView.tsx": {
+  "src/app/(dashboard)/guardrails-monitor/_components/GuardrailsMonitorView.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/guardrails-monitor/components/GuardrailsOverview.tsx": {
+  "src/app/(dashboard)/guardrails-monitor/_components/GuardrailsOverview.tsx": {
     "no-nested-ternary": {
       "count": 8
     }
   },
-  "src/app/(dashboard)/guardrails-monitor/components/ScoreChart.test.tsx": {
+  "src/app/(dashboard)/guardrails-monitor/_components/ScoreChart.test.tsx": {
     "react/display-name": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/guardrails-monitor/components/ScoreChart.tsx": {
+  "src/app/(dashboard)/guardrails-monitor/_components/ScoreChart.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
@@ -522,7 +522,7 @@
       "count": 1
     }
   },
-  "src/app/(dashboard)/projects/components/ProjectDetailsPage.tsx": {
+  "src/app/(dashboard)/projects/_components/ProjectDetailsPage.tsx": {
     "no-nested-ternary": {
       "count": 3
     },
@@ -530,17 +530,17 @@
       "count": 1
     }
   },
-  "src/app/(dashboard)/projects/components/ProjectKeysSection.tsx": {
+  "src/app/(dashboard)/projects/_components/ProjectKeysSection.tsx": {
     "react-hooks/set-state-in-effect": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/projects/components/ProjectModals/ProjectBaseForm.tsx": {
+  "src/app/(dashboard)/projects/_components/ProjectModals/ProjectBaseForm.tsx": {
     "react-hooks/set-state-in-effect": {
       "count": 2
     }
   },
-  "src/app/(dashboard)/projects/components/ProjectsPage.tsx": {
+  "src/app/(dashboard)/projects/_components/ProjectsPage.tsx": {
     "react-hooks/set-state-in-effect": {
       "count": 1
     }
