+++ ui/litellm-dashboard/src/app/(dashboard)/caching/_components/cache_health.tsx
