+++ ui/litellm-dashboard/src/app/(dashboard)/projects/_components/ProjectDetailsPage.tsx
